refactor: Split lib.rs into smaller files to ease testing, extension, and review
