+++ src/trait_impls/interactive_clap.rs
@@ -1,0 +1,5 @@
+use crate::NearGas;
+
+impl interactive_clap::ToCli for NearGas {
+    type CliVariant = NearGas;
+}
