+++ src/trait_impls/mod.rs
@@ -1,0 +1,10 @@
+#[cfg(feature = "borsh")]
+mod borsh;
+mod display;
+mod from_str;
+#[cfg(feature = "interactive-clap")]
+mod interactive_clap;
+#[cfg(feature = "schemars")]
+mod schemars;
+#[cfg(feature = "serde")]
+mod serde;
