+++ src/trait_impls/schemars.rs
@@ -1,0 +1,15 @@
+use crate::NearGas;
+
+impl schemars::JsonSchema for NearGas {
+    fn is_referenceable() -> bool {
+        false
+    }
+
+    fn schema_name() -> String {
+        String::schema_name()
+    }
+
+    fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
+        String::json_schema(gen)
+    }
+}
